git-svn-id: https://svn.intern.artcom.de/pro60/trunk/src@20152 f3027e6b-e3f5-0310-a8e1-edde99107618

+++ Y60/acgtk/ConsoleView.cpp
@@ -50,7 +50,7 @@ ConsoleView::append(const Glib::ustring & theText, const Glib::ustring & theTagN
         Gtk::TextIter myEnd = myBuffer->end();
         scroll_to( myEnd, 0.0);
 
-        while ( myBuffer->get_line_count() > _myScrollback) {
+        while ( myBuffer->get_line_count() > static_cast<int>(_myScrollback) ) {
             myBuffer->erase( myBuffer->begin(), myBuffer->get_iter_at_line( 1 ));
         }
     }
