+++ asl/net/UDPSocket.cpp
@@ -112,7 +112,7 @@ namespace inet {
 
         int byteswritten;
         AC_TRACE << "Sending to " << hex << _myRemoteEndpoint.sin_addr.s_addr << dec << ":" << _myRemoteEndpoint.sin_port;
-        if ((byteswritten=sendto(fd, (char*)data, len, 0, (struct sockaddr*)&_myRemoteEndpoint, sizeof(_myRemoteEndpoint)))!=len) {
+        if ((byteswritten=sendto(fd, (char*)data, len, 0, (struct sockaddr*)&_myRemoteEndpoint, sizeof(_myRemoteEndpoint)))!=static_cast<int>(len)) {
             throw SocketException("UDPSocket::SendTo failed");
         }
         return byteswritten;
